extract Tabs to own component and clean UI
