+++ src/components/pages/Roles/RolesDetailsDrawer.tsx
@@ -1,5 +1,4 @@
 import {
-  Box,
   Drawer,
   DrawerBody,
   DrawerContent,
@@ -9,21 +8,14 @@ import {
   GridItem,
   Icon,
   IconButton,
-  Tab,
-  TabList,
-  TabPanel,
-  TabPanels,
-  Tabs,
   Text,
 } from '@chakra-ui/react';
 import { List, PencilLine, User, X } from '@phosphor-icons/react';
 import { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Address } from 'viem';
 import { BACKGROUND_SEMI_TRANSPARENT } from '../../../constants/common';
 import useAddress from '../../../hooks/utils/useAddress';
 import useAvatar from '../../../hooks/utils/useAvatar';
-import { useCanUserCreateProposal } from '../../../hooks/utils/useCanUserSubmitProposal';
 import { useGetAccountName } from '../../../hooks/utils/useGetAccountName';
 import { useFractal } from '../../../providers/App/AppProvider';
 import {
@@ -34,42 +26,8 @@ import {
 import { BarLoader } from '../../ui/loaders/BarLoader';
 import Avatar from '../../ui/page/Header/Avatar';
 import Divider from '../../ui/utils/Divider';
-import { RolePaymentDetails } from './RolePaymentDetails';
-import RoleTermDetails from './RoleTermDetails';
-import { RoleDetailsDrawerProps, SablierPayment } from './types';
-
-function NoDataCard({
-  emptyText,
-  emptyTextNotProposer,
-}: {
-  emptyText: string;
-  emptyTextNotProposer?: string;
-}) {
-  const { t } = useTranslation(['roles']);
-  const { canUserCreateProposal } = useCanUserCreateProposal();
-  return (
-    <Box
-      bg="neutral-2"
-      boxShadow="0px 0px 0px 1px #100414, inset 0px 0px 0px 1px rgba(248, 244, 252, 0.04), inset 0px 1px 0px rgba(248, 244, 252, 0.04)"
-      borderRadius="0.75rem"
-      p="1rem"
-    >
-      <Text
-        textStyle="body-base"
-        textAlign="center"
-        color="neutral-6"
-      >
-        {t(
-          emptyTextNotProposer
-            ? canUserCreateProposal
-              ? emptyText
-              : emptyTextNotProposer
-            : emptyText,
-        )}
-      </Text>
-    </Box>
-  );
-}
+import RoleDetailsTabs from './RoleDetailsTabs';
+import { RoleDetailsDrawerProps } from './types';
 
 function RoleAndDescriptionLabel({ label, icon }: { label: string; icon: React.ElementType }) {
   return (
@@ -85,104 +43,6 @@ function RoleAndDescriptionLabel({ label, icon }: { label: string; icon: React.E
         {label}
       </Text>
     </Flex>
-  );
-}
-
-function RolesDetailsPayments({
-  payments,
-  roleHatSmartAddress,
-  roleHatWearerAddress,
-  roleTerms,
-}: {
-  payments: (Omit<SablierPayment, 'contractAddress' | 'streamId'> & {
-    contractAddress?: Address;
-    streamId?: string;
-  })[];
-  roleHatWearerAddress: Address | undefined;
-  roleHatSmartAddress: Address | undefined;
-  roleTerms: {
-    termEndDate: Date;
-    termNumber: number;
-  }[];
-}) {
-  const { t } = useTranslation(['roles']);
-  const sortedPayments = useMemo(
-    () =>
-      payments
-        ? [...payments]
-            .sort(paymentSorterByWithdrawAmount)
-            .sort(paymentSorterByStartDate)
-            .sort(paymentSorterByActiveStatus)
-        : [],
-    [payments],
-  );
-
-  if (!sortedPayments.length) {
-    return (
-      <NoDataCard
-        emptyText="noActivePayments"
-        emptyTextNotProposer="noActivePaymentsNotProposer"
-      />
-    );
-  }
-
-  return (
-    <>
-      <Divider
-        variant="darker"
-        my={4}
-      />
-      <Text
-        textStyle="display-lg"
-        color="white-0"
-      >
-        {t('payments')}
-      </Text>
-      {sortedPayments.map((payment, index) => (
-        <RolePaymentDetails
-          key={index}
-          payment={payment}
-          roleHatSmartAddress={roleHatSmartAddress}
-          roleTerms={roleTerms}
-          roleHatWearerAddress={roleHatWearerAddress}
-          showWithdraw
-        />
-      ))}
-    </>
-  );
-}
-
-function RolesDetailsTerms({
-  currentTerm,
-  nextTerm,
-  expiredTerms,
-}: {
-  nextTerm:
-    | {
-        termEndDate: Date;
-        termNumber: number;
-        nominee: string;
-      }
-    | undefined;
-  currentTerm:
-    | {
-        termEndDate: Date;
-        termNumber: number;
-        nominee: string;
-      }
-    | undefined;
-  expiredTerms: {
-    termEndDate: Date;
-    termNumber: number;
-    nominee: string;
-  }[];
-}) {
-  return (
-    <RoleTermDetails
-      currentTerm={currentTerm}
-      nextTerm={nextTerm}
-      expiredTerms={expiredTerms}
-    />
   );
 }
 
@@ -321,39 +181,12 @@ export default function RolesDetailsDrawer({
             variant="darker"
             my={4}
           />
-          <Tabs
-            variant="twoTone"
-            mt={4}
-          >
-            <TabList>
-              <Tab>{t('terms')}</Tab>
-              <Tab>{t('payments')}</Tab>
-            </TabList>
-            <TabPanels mt={4}>
-              <TabPanel>
-                <RolesDetailsTerms
-                  currentTerm={roleHat.roleTerms.currentTerm}
-                  nextTerm={roleHat.roleTerms.nextTerm}
-                  expiredTerms={roleHat.roleTerms.expiredTerms.map(term => ({
-                    nominee: term.nominee,
-                    termEndDate: term.termEndDate,
-                    termNumber: term.termNumber,
-                  }))}
-                />
-              </TabPanel>
-              <TabPanel>
-                <RolesDetailsPayments
-                  payments={sortedPayments}
-                  roleTerms={roleHat.roleTerms.allTerms.map(term => ({
-                    termEndDate: term.termEndDate,
-                    termNumber: term.termNumber,
-                  }))}
-                  roleHatSmartAddress={roleHat.smartAddress}
-                  roleHatWearerAddress={roleHatWearerAddress}
-                />
-              </TabPanel>
-            </TabPanels>
-          </Tabs>
+          <RoleDetailsTabs
+            roleHatSmartAddress={roleHat.smartAddress}
+            roleTerms={roleHat.roleTerms}
+            roleHatWearerAddress={roleHatWearerAddress}
+            sortedPayments={sortedPayments}
+          />
         </DrawerBody>
       </DrawerContent>
     </Drawer>
