+++ src/components/pages/Roles/RolesDetailsDrawerMobile.tsx
@@ -11,9 +11,8 @@ import {
 } from '../../../store/roles/rolesStoreUtils';
 import { useRolesStore } from '../../../store/roles/useRolesStore';
 import DraggableDrawer from '../../ui/containers/DraggableDrawer';
-import Divider from '../../ui/utils/Divider';
 import { AvatarAndRoleName } from './RoleCard';
-import { RolePaymentDetails } from './RolePaymentDetails';
+import RoleDetailsTabs from './RoleDetailsTabs';
 import { RoleDetailsDrawerProps } from './types';
 
 export default function RolesDetailsDrawerMobile({
@@ -98,35 +97,12 @@ export default function RolesDetailsDrawerMobile({
         px="1rem"
         mb="1.5rem"
       >
-        {roleHat.payments && (
-          <>
-            <Divider
-              variant="darker"
-              my={4}
-            />
-            <Text
-              textStyle="display-lg"
-              color="white-0"
-              mt="1.5rem"
-              mb="1rem"
-            >
-              {t('payments')}
-            </Text>
-            {sortedPayments.map((payment, index) => (
-              <RolePaymentDetails
-                key={index}
-                payment={payment}
-                roleHatSmartAddress={roleHat.smartAddress}
-                roleHatWearerAddress={roleHatWearerAddress}
-                roleTerms={roleHat.roleTerms.allTerms.map(term => ({
-                  termEndDate: term.termEndDate,
-                  termNumber: term.termNumber,
-                }))}
-                showWithdraw
-              />
-            ))}
-          </>
-        )}
+        <RoleDetailsTabs
+          roleHatSmartAddress={roleHat.smartAddress}
+          roleTerms={roleHat.roleTerms}
+          roleHatWearerAddress={roleHatWearerAddress}
+          sortedPayments={sortedPayments}
+        />
       </Box>
     </DraggableDrawer>
   );
